rewrite!
Passing around adjacency matrices or lists of edges, and converting
between them all the time, was awkward.

The extension now defines data types for vertices, edges, and graphs.

There are functions to show the working-out for Kruskal's and Prim's
algorithms. (and a nice framework for building up working-out steps in
the course of an algorithm, which I might transfer to the main Numbas
runtime)